requirements.txt
numpy
pandas
scipy
torch
matplotlib
torchvision
Pillow
